refactor:  chnage for Angular 14

+++ client/src/app/modules/users/pages/user-edit/user-edit.component.ts
@@ -1,13 +1,10 @@
 
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
-import { OpenMode } from 'src/app/__shared__/model/enums';
-import { IUser } from 'src/app/__shared__/model/user';
-import { MessageService } from 'src/app/__shared__/modules/messages/message-service/message.service';
+import { IUser, OpenMode } from 'src/app/__shared__/model';
+import { MessageService } from 'src/app/__shared__/modules/messages';
 import { AuthService } from '../../../access/services/auth.service';
 import { UserService } from '../../services/user-service/user-service';
-
-
 
 @Component({
   selector: 'user-edit',
@@ -16,18 +13,17 @@ import { UserService } from '../../services/user-service/user-service';
 })
 export class UserEditComponent implements OnInit {
   mode: OpenMode = OpenMode.ADD;
-  user: IUser = {} as IUser;
+  user = {} as IUser;
 
   constructor(
     private authService: AuthService,
     private userService: UserService,
     private router: Router,
     private route: ActivatedRoute,
-    private messageService: MessageService
-
+    private messageService: MessageService,
   ) { }
 
-  ngOnInit() {
+  ngOnInit(): void {
     const id: string = this.route.snapshot.params['id'];
 
     if (id) {
@@ -39,7 +35,7 @@ export class UserEditComponent implements OnInit {
     }
   }
 
-  loadUser(id: string) {
+  loadUser(id: string): void {
     if (this.authService.currentUser) {
       this.user = this.authService.currentUser;
     }
@@ -57,10 +53,10 @@ export class UserEditComponent implements OnInit {
   }
 
   save() {
-    this.userService
+    this.user && this.userService
       .saveUser(this.user)
       .subscribe(
-        () => this.onSaveCompleted(`${this.user.fullName} was saved`)
+        () => this.onSaveCompleted(`${this.user?.fullName} was saved`)
       );
   }
 
@@ -72,13 +68,13 @@ export class UserEditComponent implements OnInit {
     this.router.navigate(['/users']);
   }
 
-  reset() {
+  reset(): void {
     // TODO = REFACTORING
     // console.warn("reset()  - Method not implemented.");
   }
 
-  delete() {
-    this.userService.removeUser(this.user.id)
+  delete(): void {
+    this.user && this.userService.removeUser(this.user.id)
       .subscribe(_ =>
         this.router.navigate(['/users'])
       );
